Move test to the right place
