+++ subprojects/core/src/test/groovy/org/gradle/internal/operations/DefaultBuildOperationExecutorTest.groovy
@@ -14,23 +14,12 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.progress
+package org.gradle.internal.operations
 
 import org.gradle.internal.concurrent.DefaultParallelismConfiguration
 import org.gradle.internal.concurrent.ExecutorFactory
 import org.gradle.internal.concurrent.GradleThread
-import org.gradle.internal.operations.BuildOperationContext
-import org.gradle.internal.operations.BuildOperationDescriptor
-import org.gradle.internal.operations.BuildOperationListener
-import org.gradle.internal.operations.BuildOperationQueueFactory
-import org.gradle.internal.operations.BuildOperationRef
-import org.gradle.internal.operations.CallableBuildOperation
-import org.gradle.internal.operations.CurrentBuildOperationRef
-import org.gradle.internal.operations.DefaultBuildOperationExecutor
-import org.gradle.internal.operations.DefaultBuildOperationIdFactory
-import org.gradle.internal.operations.OperationFinishEvent
-import org.gradle.internal.operations.OperationStartEvent
-import org.gradle.internal.operations.RunnableBuildOperation
+import org.gradle.internal.progress.NoOpProgressLoggerFactory
 import org.gradle.internal.time.Clock
 import org.gradle.test.fixtures.concurrent.ConcurrentSpec
